Attempt at adding Exp. Candies
To do:
-Fix bigger candies not giving multiple levels when possible.

+++ include/item.h
@@ -21,7 +21,7 @@ struct Item
     ItemUseFunc fieldUseFunc;
     u8 battleUsage;
     ItemUseFunc battleUseFunc;
-    u8 secondaryId;
+    u16 secondaryId;
     u8 flingPower;
 };
